Cleanup and refactor 2
