+++ package.json
@@ -156,7 +156,7 @@
     "lint-scss-fix": "stylelint src/**/*.scss --syntax scss --fix",
     "precommit": "lint-staged",
     "start": "run-s start-ts",
-    "start-ts": "ts-node --files scripts/serve.ts",
+    "start-ts": "ts-node --files scripts/start.ts",
     "start-story": "start-storybook -p 9001 -c .storybook",
     "test": "jest --watch --env=jsdom",
     "test-coverage": "jest --coverage --env=jsdom"
